implementando pagamento online com stripe

+++ Domain/Entities/Plan.cs
@@ -9,7 +9,7 @@ namespace Domain.Entities
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public required Double Value { get; set; }
+        public required Decimal Value { get; set; }
         public required int Duration { get; set; }
         public bool Active { get; set; } = true;
